Added sampleproviders for mixing (may need debugging)

+++ Beater/Beater.Shared/ViewModels/PatternViewModel.cs
@@ -21,8 +21,8 @@ namespace Beater.ViewModels
             Beats = new ObservableCollection<BeatViewModel>(BeatViewModel.BeatPattern(beat));
         }
 
-        public long BeatLength { get { return beat.BeatLength; } set { beat.BeatLength = value; } }
-        public long Measure { get { return beat.Measure; } set { beat.Measure = value; } }
+        public int BeatLength { get { return beat.BeatLength; } set { beat.BeatLength = value; } }
+        public int Measure { get { return beat.Measure; } set { beat.Measure = value; } }
         public ObservableCollection<BeatViewModel> Beats { get; private set; }
         public string Id { get { return beat.Id; } set { beat.Id = value; } }
     }
